fix(ProjectTimelineLabel): Final $el fix
When a project has a end which is smaller than the start, the component will
not display. mounted() will still be called and executes getComputedStyle with
this.$el as an argument. In this specific case, undefined will be passed into
the function and an error will be thrown.

Fix is to check if this.$el is available and only executing the function
if there is a rendered element.